+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/HierarchyEntity.java
@@ -2,7 +2,9 @@ package uk.ac.bbsrc.tgac.miso.core.data;
 
 import java.math.BigDecimal;
 
-public interface HierarchyEntity extends Boxable, ChangeLoggable {
+import uk.ac.bbsrc.tgac.miso.core.data.qc.DetailedQcItem;
+
+public interface HierarchyEntity extends Boxable, ChangeLoggable, DetailedQcItem {
 
   public HierarchyEntity getParent();