+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/DetailedLibraryAliquotDto.java
@@ -7,8 +7,12 @@ public class DetailedLibraryAliquotDto extends LibraryAliquotDto {
 
   private boolean nonStandardAlias;
   private Long libraryDesignCodeId;
+  private String identityConsentLevel;
+  private String effectiveExternalNames;
   private String effectiveGroupId;
   private String effectiveGroupIdSample;
+  private String effectiveTissueOriginLabel;
+  private String effectiveTissueTypeLabel;
   private String groupId;
   private String groupDescription;
 
@@ -28,6 +32,22 @@ public class DetailedLibraryAliquotDto extends LibraryAliquotDto {
     this.libraryDesignCodeId = libraryDesignCodeId;
   }
 
+  public String getIdentityConsentLevel() {
+    return identityConsentLevel;
+  }
+
+  public void setIdentityConsentLevel(String identityConsentLevel) {
+    this.identityConsentLevel = identityConsentLevel;
+  }
+
+  public String getEffectiveExternalNames() {
+    return effectiveExternalNames;
+  }
+
+  public void setEffectiveExternalNames(String effectiveExternalNames) {
+    this.effectiveExternalNames = effectiveExternalNames;
+  }
+
   public String getEffectiveGroupId() {
     return effectiveGroupId;
   }
@@ -42,6 +62,22 @@ public class DetailedLibraryAliquotDto extends LibraryAliquotDto {
 
   public void setEffectiveGroupIdSample(String effectiveGroupIdSample) {
     this.effectiveGroupIdSample = effectiveGroupIdSample;
+  }
+
+  public String getEffectiveTissueOriginLabel() {
+    return effectiveTissueOriginLabel;
+  }
+
+  public void setEffectiveTissueOriginLabel(String effectiveTissueOriginLabel) {
+    this.effectiveTissueOriginLabel = effectiveTissueOriginLabel;
+  }
+
+  public String getEffectiveTissueTypeLabel() {
+    return effectiveTissueTypeLabel;
+  }
+
+  public void setEffectiveTissueTypeLabel(String effectiveTissueTypeLabel) {
+    this.effectiveTissueTypeLabel = effectiveTissueTypeLabel;
   }
 
   public String getGroupId() {