+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultLibraryService.java
@@ -551,6 +551,7 @@ public class DefaultLibraryService implements LibraryService, PaginatedDataSourc
     validateVolumeUnits(library.getVolume(), library.getVolumeUnits(), errors);
     validateBarcodeUniqueness(library, beforeChange, libraryDao::getByBarcode, errors, "library");
     validateUnboxableFields(library, errors);
+    validateDetailedQcStatus(library, errors);
     if (isDetailedLibrary(library) && beforeChange != null) {
       validateTargetedSequencing(((DetailedLibrary) library).getLibraryDesignCode(), beforeChange.getLibraryAliquots(), errors);
     }